+++ src/main/java/ma/vi/esql/database/SqlServer.java
@@ -26,8 +26,6 @@ import static ma.vi.esql.type.BaseType.BASE_TYPE;
 public class SqlServer extends AbstractDatabase {
   public SqlServer(Map<String, Object> config,
                    boolean createCoreTables) {
-    this.config = config;
-
     Properties props = new Properties();
     props.setProperty("dataSourceClassName", SQLServerConnectionPoolDataSource.class.getName());
     props.setProperty("dataSource.serverName", valueOf(config.getOrDefault("database.host", "localhost")));
@@ -41,12 +39,7 @@ public class SqlServer extends AbstractDatabase {
     dataSource = new HikariDataSource(new HikariConfig(props));
 
     init(config);
-    postInit(pooledConnection(), structure(), createCoreTables);
-  }
-
-  @Override
-  public Map<String, Object> config() {
-    return config;
+    postInit(pooledConnection(), structure());
   }
 
   @Override
@@ -55,17 +48,15 @@ public class SqlServer extends AbstractDatabase {
   }
 
   @Override
-  public void postInit(Connection con,
-                       Structure structure,
-                       boolean createCoreTables) {
-    super.postInit(con, structure, createCoreTables);
+  public void postInit(Connection con, Structure structure) {
+    super.postInit(con, structure);
     try (Connection c = pooledConnection(true, -1)) {
 
       // MS Sql-server specific initialization
 
       // Allow snapshot isolation
       c.createStatement().executeUpdate(
-          "alter database " + valueOf(config.get("database.name")) + " set allow_snapshot_isolation on");
+          "alter database " + valueOf(config().get("database.name")) + " set allow_snapshot_isolation on");
 
       // function to add intervals to dates
       c.createStatement().executeUpdate(
@@ -232,191 +223,6 @@ public class SqlServer extends AbstractDatabase {
               "  return @Interval;\n" +
               "end;\n");
 
-      // @todo: move to platform project
-//      // function to find value from lookups
-//      c.createStatement().executeUpdate(
-//          "create or alter function _core.lookup_label0(@Code nvarchar(max),\n" +
-//              "                                             @Lookup nvarchar(max),\n" +
-//              "                                             @ShowCode bit,\n" +
-//              "                                             @ShowLabel bit) returns nvarchar(max) as\n" +
-//              "begin\n" +
-//              "  declare @Result nvarchar(max);\n" +
-//              "  select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),\n" +
-//              "                      code + ' - '+ label,\n" +
-//              "                      iif(coalesce(@ShowCode, 0)=1,\n" +
-//              "                          code,\n" +
-//              "                          label)))\n" +
-//              "    from \"_platform.lookup\".\"LookupValue\" v\n" +
-//              "    join \"_platform.lookup\".\"Lookup\" l on v.lookup_id=l._id\n" +
-//              "  where l.name=@Lookup and v.code=@Code;\n" +
-//              "  return @Result;\n" +
-//              "end;");
-//
-//      c.createStatement().executeUpdate("create or alter function _core.lookup_label1(@Code nvarchar(max),\n" +
-//                                            "                                             @Lookup nvarchar(max),\n" +
-//                                            "                                             @ShowCode bit,\n" +
-//                                            "                                             @ShowLabel bit,\n" +
-//                                            "                                             @Link1 nvarchar(max)) returns nvarchar(max) as\n" +
-//                                            "begin\n" +
-//                                            "  declare @LinkCursor Cursor;\n" +
-//                                            "  declare @Result nvarchar(max);\n" +
-//                                            "\n" +
-//                                            "  select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),\n" +
-//                                            "                      v1.code + ' - '+ v1.label,\n" +
-//                                            "                      iif(coalesce(@ShowCode, 0)=1,\n" +
-//                                            "                          v1.code,\n" +
-//                                            "                          v1.label)))\n" +
-//                                            "\n" +
-//                                            "    from \"_platform.lookup\".\"LookupValue\" v0\n" +
-//                                            "    join \"_platform.lookup\".\"Lookup\" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk0 on (v0._id=lk0.source_lookup_value_id and lk0.lookup_link=@Link1)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v1 on v1._id=lk0.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "   where v0.code=@Code;\n" +
-//                                            "  return @Result;\n" +
-//                                            "end;");
-//
-//      c.createStatement().executeUpdate("create or alter function _core.lookup_label2(@Code nvarchar(max),\n" +
-//                                            "                                             @Lookup nvarchar(max),\n" +
-//                                            "                                             @ShowCode bit,\n" +
-//                                            "                                             @ShowLabel bit,\n" +
-//                                            "                                             @Link1 nvarchar(max),\n" +
-//                                            "                                             @Link2 nvarchar(max)) returns nvarchar(max) as\n" +
-//                                            "begin\n" +
-//                                            "  declare @LinkCursor Cursor;\n" +
-//                                            "  declare @Result nvarchar(max);\n" +
-//                                            "\n" +
-//                                            "  select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),\n" +
-//                                            "                      v2.code + ' - '+ v2.label,\n" +
-//                                            "                      iif(coalesce(@ShowCode, 0)=1,\n" +
-//                                            "                          v2.code,\n" +
-//                                            "                          v2.label)))\n" +
-//                                            "\n" +
-//                                            "    from \"_platform.lookup\".\"LookupValue\" v0\n" +
-//                                            "    join \"_platform.lookup\".\"Lookup\" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk0 on (v0._id=lk0.source_lookup_value_id and lk0.lookup_link=@Link1)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v1 on v1._id=lk0.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk1 on (v1._id=lk1.source_lookup_value_id and lk1.lookup_link=@Link2)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v2 on v2._id=lk1.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "   where v0.code=@Code;\n" +
-//                                            "  return @Result;\n" +
-//                                            "end;");
-//
-//      c.createStatement().executeUpdate("create or alter function _core.lookup_label3(@Code nvarchar(max),\n" +
-//                                            "                                             @Lookup nvarchar(max),\n" +
-//                                            "                                             @ShowCode bit,\n" +
-//                                            "                                             @ShowLabel bit,\n" +
-//                                            "                                             @Link1 nvarchar(max),\n" +
-//                                            "                                             @Link2 nvarchar(max),\n" +
-//                                            "                                             @Link3 nvarchar(max)) returns nvarchar(max) as\n" +
-//                                            "begin\n" +
-//                                            "  declare @LinkCursor Cursor;\n" +
-//                                            "  declare @Result nvarchar(max);\n" +
-//                                            "\n" +
-//                                            "  select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),\n" +
-//                                            "                      v3.code + ' - '+ v3.label,\n" +
-//                                            "                      iif(coalesce(@ShowCode, 0)=1,\n" +
-//                                            "                          v3.code,\n" +
-//                                            "                          v3.label)))\n" +
-//                                            "\n" +
-//                                            "    from \"_platform.lookup\".\"LookupValue\" v0\n" +
-//                                            "    join \"_platform.lookup\".\"Lookup\" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk0 on (v0._id=lk0.source_lookup_value_id and lk0.lookup_link=@Link1)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v1 on v1._id=lk0.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk1 on (v1._id=lk1.source_lookup_value_id and lk1.lookup_link=@Link2)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v2 on v2._id=lk1.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk2 on (v2._id=lk2.source_lookup_value_id and lk2.lookup_link=@Link3)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v3 on v3._id=lk2.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "   where v0.code=@Code;\n" +
-//                                            "  return @Result;\n" +
-//                                            "end;");
-//
-//      c.createStatement().executeUpdate("create or alter function _core.lookup_label4(@Code nvarchar(max),\n" +
-//                                            "                                             @Lookup nvarchar(max),\n" +
-//                                            "                                             @ShowCode bit,\n" +
-//                                            "                                             @ShowLabel bit,\n" +
-//                                            "                                             @Link1 nvarchar(max),\n" +
-//                                            "                                             @Link2 nvarchar(max),\n" +
-//                                            "                                             @Link3 nvarchar(max),\n" +
-//                                            "                                             @Link4 nvarchar(max)) returns nvarchar(max) as\n" +
-//                                            "begin\n" +
-//                                            "  declare @LinkCursor Cursor;\n" +
-//                                            "  declare @Result nvarchar(max);\n" +
-//                                            "\n" +
-//                                            "  select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),\n" +
-//                                            "                      v4.code + ' - '+ v4.label,\n" +
-//                                            "                      iif(coalesce(@ShowCode, 0)=1,\n" +
-//                                            "                          v4.code,\n" +
-//                                            "                          v4.label)))\n" +
-//                                            "\n" +
-//                                            "    from \"_platform.lookup\".\"LookupValue\" v0\n" +
-//                                            "    join \"_platform.lookup\".\"Lookup\" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk0 on (v0._id=lk0.source_lookup_value_id and lk0.lookup_link=@Link1)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v1 on v1._id=lk0.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk1 on (v1._id=lk1.source_lookup_value_id and lk1.lookup_link=@Link2)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v2 on v2._id=lk1.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk2 on (v2._id=lk2.source_lookup_value_id and lk2.lookup_link=@Link3)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v3 on v3._id=lk2.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk3 on (v3._id=lk3.source_lookup_value_id and lk3.lookup_link=@Link4)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v4 on v4._id=lk3.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "   where v0.code=@Code;\n" +
-//                                            "  return @Result;\n" +
-//                                            "end;");
-//
-//      c.createStatement().executeUpdate("create or alter function _core.lookup_label5(@Code nvarchar(max),\n" +
-//                                            "                                             @Lookup nvarchar(max),\n" +
-//                                            "                                             @ShowCode bit,\n" +
-//                                            "                                             @ShowLabel bit,\n" +
-//                                            "                                             @Link1 nvarchar(max),\n" +
-//                                            "                                             @Link2 nvarchar(max),\n" +
-//                                            "                                             @Link3 nvarchar(max),\n" +
-//                                            "                                             @Link4 nvarchar(max),\n" +
-//                                            "                                             @Link5 nvarchar(max)) returns nvarchar(max) as\n" +
-//                                            "begin\n" +
-//                                            "  declare @LinkCursor Cursor;\n" +
-//                                            "  declare @Result nvarchar(max);\n" +
-//                                            "\n" +
-//                                            "  select @Result=(iif(coalesce(@ShowCode, 0)=coalesce(@ShowLabel, 0),\n" +
-//                                            "                      v5.code + ' - '+ v5.label,\n" +
-//                                            "                      iif(coalesce(@ShowCode, 0)=1,\n" +
-//                                            "                          v5.code,\n" +
-//                                            "                          v5.label)))\n" +
-//                                            "\n" +
-//                                            "    from \"_platform.lookup\".\"LookupValue\" v0\n" +
-//                                            "    join \"_platform.lookup\".\"Lookup\" lookup on (v0.lookup_id=lookup._id and lookup.name=@Lookup)\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk0 on (v0._id=lk0.source_lookup_value_id and lk0.lookup_link=@Link1)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v1 on v1._id=lk0.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk1 on (v1._id=lk1.source_lookup_value_id and lk1.lookup_link=@Link2)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v2 on v2._id=lk1.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk2 on (v2._id=lk2.source_lookup_value_id and lk2.lookup_link=@Link3)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v3 on v3._id=lk2.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk3 on (v3._id=lk3.source_lookup_value_id and lk3.lookup_link=@Link4)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v4 on v4._id=lk3.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValueLink\" lk4 on (v4._id=lk4.source_lookup_value_id and lk4.lookup_link=@Link5)\n" +
-//                                            "    join \"_platform.lookup\".\"LookupValue\" v5 on v5._id=lk4.target_lookup_value_id\n" +
-//                                            "\n" +
-//                                            "   where v0.code=@Code;\n" +
-//                                            "  return @Result;\n" +
-//                                            "end;\n");
-
       c.createStatement().executeUpdate("create or alter function _core.rpad(@Str nvarchar(max), @Length int, @Pad nvarchar=' ') returns nvarchar(max)\n" +
                                             "begin\n" +
                                             "  declare @Remaining int;\n" +
@@ -576,11 +382,11 @@ public class SqlServer extends AbstractDatabase {
     try {
       Connection con = DriverManager.getConnection(
           "jdbc:sqlserver://"
-              + valueOf(config.get("database.host"))
-              + (config.containsKey("database.port")
-                    ? ':' + valueOf(config.get("database.port"))
+              + valueOf(config().get("database.host"))
+              + (config().containsKey("database.port")
+                    ? ':' + valueOf(config().get("database.port"))
                     : "")
-              + ";database=" + valueOf(config.get("database.name"))
+              + ";database=" + valueOf(config().get("database.name"))
               + ";sendStringParametersAsUnicode=true",
           username,
           password);
@@ -695,6 +501,4 @@ public class SqlServer extends AbstractDatabase {
    * Datasource for pooled connections.
    */
   private final HikariDataSource dataSource;
-
-  private final Map<String, Object> config;
 }